facial_expressions_classifier_project/requirements.txt
torch
torchvision
flask
flask-cors
pillow
matplotlib